Allow backups against `archive_mode = always` standbys to complete
Changes the way backups are stopped on standby servers by concurrently
calling pg_switch_wal() on the primary during the pg_backup_stop call.

This is currently a proof-of-concept which works for rsync-concurrent
and rsync-exclusive backups only. For postgres backups there are
additional challenges, since in those cases we are not calling
pg_backup_stop ourselves.

This implementation uses a child process to perform the WAL switch
on the primary before calling the same function we would normally use
to stop a concurrent backup. This child process will perform a set
number of WAL switches at a set number of intervals unless the parent
process completes stopping the backup and stops the child process by
sending a `True` value to its done queue.

This is far from perfect and may not *always* allow the backup to
complete. Ideally we would spawn the pg_backup_stop command in a child
process, wait until we knew it was waiting for the WAL switch, and then
perform the WAL switch either in the parent or another child process.
However there are several reasons this commit does not do that:

1. The pg_backup_stop and pg_backup_start functions must be called
   using the same connection and connections cannot be shared across
   processes. We therefore cannot perform the pg_backup_stop command
   in a child process unless we also perform the pg_backup_start
   command in the same child process. This would some extra plumbing
   in the backup executor which would increase the complexity of the
   change.
2. Even if we run pg_backup_stop in a child process there is no
   obvious way of determining that it is actually waiting for the
   WAL switch. We would still therefore have to wait an arbitrary
   amount of time before switching WALs, and then potentially perform
   subsequent WAL switches in the case that the first switches were
   still too early.

One option which wasn't considered here was to use threads instead of
processes - this would allow the PostgreSQL connection to be shared
(as long as they do not share cursors psycopg2 is safe). This may
require less plumbing overall than if using multiprocessing, since
the pg_backup_start call could happen in the main thread. We'd still
be left with the second problem however, so we'd ultimately end up
doing something like this commit anyway, just with more plumbing.

Overall I think the trade-offs of the approach here are acceptable.
For the "my primary is too quiet to produce a WAL switch" scenarios
the chances are the first WAL switch will be enough and the stop
backup call will complete successfully. If there happens to be
some write activity between first WAL switch and the stop backup
call then perhaps a second WAL switch may occur, but generally the
stop backup call should complete in a timely fashion.

For the case where servers are busier, we would expect WAL switches
to be occurring regularly anyway so the change in this commit would
be a no-op most of the time.

In the absolute worst case where pg_backup_stop is not yet waiting
for the WAL switch and the child process is switching WALs every
`wait` interval then this is wasteful but the consequence is
limited to `n=wait` additional WALs and a backup which would have
ended up in a hanged or failed state anyway.

Relates to #596.