Fix logging handlers for pg_basebackup command
Replaces the output and error lists used as handlers when processing
command output with wrapper functions which collect output in those
lists *and* call the out_handler and err_handler functions if they
are set. By default, the out_handler and err_handler functions will
log at info and warning levels when they are called, so this change
allows these functions to log while the command is executing.

This restores the logging of commands such as pg_basebackup to the
barman output while the commands are running. This means that in
the event of pg_basebackup (or other commands) waiting for a long
time, e.g. for a WAL switch, it will be possible to see the stdout
and stderr produced so far in the barman log.

Closes #493.